+++ container-engine-lib/lib/backend_impls/docker/docker_kurtosis_backend/engine_functions/destroy_engines.go
@@ -62,16 +62,9 @@ func DestroyEngines(
 	for guidStr, err := range erroredEngineGuidStrs {
 		erroredGuids[engine.EngineGUID(guidStr)] = stacktrace.Propagate(
 			err,
-			"An error occurred destroying engine '%v'",
+		"An error occurred destroying engine '%v'",
 			guidStr,
 		)
-	}
-
-	//TODO we are removing the los components containers rather than stopping them because we are preparing the stage
-	//TODO for a single engine server, logs components containers have an static name, so if we stop the containers
-	//TODO the engine restart will fail because the container's name will be in use
-	if err := removeLogsComponentsGracefully(ctx, filters, dockerManager); err != nil {
-		return nil, nil, stacktrace.Propagate(err, "An error occurred removing the logs components containers")
 	}
 
 	return successfulGuids, erroredGuids, nil