+++ container-engine-lib/lib/backend_impls/docker/docker_kurtosis_backend/engine_functions/logs_components/logs_collectors_type.go
@@ -1,4 +1,0 @@
-package logs_components
-
-type LogsCollectorAddress string
-type LogsCollectorLabels []string