+++ docs/changelog.md
@@ -1,4 +1,14 @@
 # TBD
+### Breaking Changes
+* Updated `CreateEngine` method in `container-engine-lib`, removed the `logsCollectorHttpPortNumber` parameter
+    * Users will need to update all the `CreateEngine` calls removing this new parameter
+
+### Features
+* Created the `LogsDatabase` object in `container-engine-lib`
+* Created the `LogsCollector` object in `container-engine-lib`
+
+### Changes
+* Untied the logs components containers and volumes creation and removal from the engine's crud in `container-engine-lib`
 
 ### Features
 * Implement skeleton for the Startosis engine
@@ -44,7 +54,6 @@
 * Fixed circle to not docs check on merge
 
 # 0.49.1
-
 ### Fixes
 * Attempting to fix the release version
 ### Changes